Distinguish two pair from pair in Deuces Wild hold classification
HoldClassifier.classifyDeucesWild collapsed a held two pair into .pair
because two pair doesn't pay in Deuces Wild. That erased the strategic
distinction between holding one pair and holding two, which made
downstream coaching text (in the video poker app) unable to describe
what was actually wrong with a two-pair hold without app-side special
casing.

Non-paying hands are still valid strategic categories worth naming.
classifyDW0Four and classifyDW0Five now return .twoPair when the held
cards form two separate pairs, matching the existing Jacks or Better
behavior.

Moved the DW classifier threshold tests and added four new pair /
two-pair tests into a new HoldClassifierDeucesWildTests.swift, keeping
DeucesWildTests.swift under SwiftLint's type_body_length limit.
