+++ Tests/PlayingCardTests/DeucesWildTests.swift
@@ -376,66 +376,4 @@ final class DeucesWildTests: XCTestCase {
         XCTAssertEqual(result96.optimalEV, 6.0, accuracy: 0.001)
         XCTAssertEqual(result95.optimalEV, 5.0, accuracy: 0.001)
     }
-
-    // MARK: - Strategy pattern threshold correctness
-
-    func testDW2DeuceSFThresholdBelowSix() {
-        // 2♣ 2♦ 4♣ 5♣ + K♦: both naturals below rank 6 → should NOT be fourToStraightFlushTwoDeuces.
-        let hand = [
-            card(.two, .clubs), card(.two, .diamonds),
-            card(.four, .clubs), card(.five, .clubs), card(.king, .diamonds),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2, 3])
-        XCTAssertEqual(pattern, .twoDeuces)
-    }
-
-    func testDW2DeuceSFThresholdAtSix() {
-        // 2♣ 2♦ 6♣ 9♣ + K♦: both naturals ≥ 6, span 3, same suit → fourToStraightFlushTwoDeuces.
-        let hand = [
-            card(.two, .clubs), card(.two, .diamonds),
-            card(.six, .clubs), card(.nine, .clubs), card(.king, .diamonds),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2, 3])
-        XCTAssertEqual(pattern, .fourToStraightFlushTwoDeuces)
-    }
-
-    func testDW1DeuceSFHighThresholdBelowFive() {
-        // 2♠ 3♣ 4♣ 5♣ + K♦: consecutive naturals 3-4-5, lowest < 5 → fourToStraightFlushOneDeuce (not high).
-        let hand = [
-            card(.two, .spades), card(.three, .clubs), card(.four, .clubs),
-            card(.five, .clubs), card(.king, .diamonds),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2, 3])
-        XCTAssertEqual(pattern, .fourToStraightFlushOneDeuce)
-    }
-
-    func testDW1DeuceSFHighThresholdAtFive() {
-        // 2♠ 5♣ 6♣ 7♣ + K♦: consecutive naturals 5-6-7, lowest = 5 → fourToStraightFlushHighOneDeuce.
-        let hand = [
-            card(.two, .spades), card(.five, .clubs), card(.six, .clubs),
-            card(.seven, .clubs), card(.king, .diamonds),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2, 3])
-        XCTAssertEqual(pattern, .fourToStraightFlushHighOneDeuce)
-    }
-
-    func testDW1Deuce3SFHighThresholdBelowSix() {
-        // 2♠ 4♥ 5♥ + K♦ Q♠: consecutive naturals 4-5, lowest < 6 → oneDeuce.
-        let hand = [
-            card(.two, .spades), card(.four, .hearts), card(.five, .hearts),
-            card(.king, .diamonds), card(.queen, .spades),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2])
-        XCTAssertEqual(pattern, .oneDeuce)
-    }
-
-    func testDW1Deuce3SFHighThresholdAtSix() {
-        // 2♠ 6♥ 7♥ + K♦ Q♠: consecutive naturals 6-7, lowest = 6 → threeToStraightFlushHighOneDeuce.
-        let hand = [
-            card(.two, .spades), card(.six, .hearts), card(.seven, .hearts),
-            card(.king, .diamonds), card(.queen, .spades),
-        ]
-        let pattern = HoldClassifier.classifyDeucesWild(hand: hand, holding: [0, 1, 2])
-        XCTAssertEqual(pattern, .threeToStraightFlushHighOneDeuce)
-    }
 }
